#94: move voting API to /api endpoint
