+++ scalajvm/app/controllers/api/Quotes.scala
@@ -1,37 +1,32 @@
 package controllers.api
 
-import helpers.{ActionWithTx, Cors, RequestWithSession}
+import helpers.Cors
 import models.queries.{QuoteFilter, QuoteOrdering, QuoteQueries}
 import models.data.Quote
 import play.api.mvc._
 import ru.org.codingteam.loglist.dto.QuoteDTO
 import ru.org.codingteam.loglist.QuoteCount
 
-object Quotes extends Controller {
-  private def action(responseAction: RequestWithSession[AnyContent] => Result) = ActionWithTx { request =>
-    val response = responseAction(request)
-    val headers = Cors.headers(request.headers)
-    response.withHeaders(headers: _*)
-  }
-
-  def getQuote(id: Long) = action { request =>
+object Quotes extends Controller with Cors {
+  def getQuote(id: Long) = corsActionWithTx { request =>
     import request.dbSession
     prepareResponse(QuoteQueries().getQuoteById(id))
   }
 
-  def getRandomQuote = action { request =>
+  def getRandomQuote = corsActionWithTx { request =>
     import request.dbSession
     prepareResponse(QuoteQueries().getRandomQuote)
   }
 
-  def getCount(filter: QuoteFilter.Value) = action { request =>
-      import request.dbSession
-      val count = QuoteQueries().countQuotes(filter)
-      val response = QuoteCount(count)
-      json(upickle.write(response))
-    }
+  def getCount(filter: QuoteFilter.Value) = corsActionWithTx { request =>
+    import request.dbSession
+    val count = QuoteQueries().countQuotes(filter)
+    val response = QuoteCount(count)
+    json(upickle.write(response))
+  }
 
-  def getList(limit: Int, page: Int, order: QuoteOrdering.Value, filter: QuoteFilter.Value) = action { request =>
+  def getList(limit: Int, page: Int, order: QuoteOrdering.Value, filter: QuoteFilter.Value) =
+    corsActionWithTx { request =>
       import request.dbSession
 
       val pageSize = if (0 <= limit && limit <= 1000) limit else 50
