+++ scalajvm/app/controllers/api/Voting.scala
@@ -1,14 +1,12 @@
-package controllers
+package controllers.api
 
-import helpers.ActionWithTx
+import helpers.Cors
 import models.queries.VotingQueries
-import play.api.Play.current
 import play.api.mvc._
-import scalikejdbc._
 import ru.org.codingteam.loglist.QuoteRating
 
-object Voting extends Controller {
-  def vote(id: Long, up: Boolean) = ActionWithTx { request =>
+object Voting extends Controller with Cors {
+  def vote(id: Long, up: Boolean) = corsActionWithTx { request =>
     import request.dbSession
 
     val rating = VotingQueries().updateRating(if (up) 1 else -1)(id)
