+++ scalajvm/app/global/Options.scala
@@ -3,8 +3,8 @@ package global
 import helpers.Cors
 import play.api.mvc._
 
-object Options extends Controller {
+object Options extends Controller with Cors {
   def corsSupport(url: String) = Action { request =>
-    NoContent.withHeaders(Cors.headers(request.headers): _*)
+    NoContent.withHeaders(corsHeaders(request.headers): _*)
   }
 }
